fix(autoware_motion_utils): fix bugprone-narrowing-conversions warnings (#932)
* fix(autoware_motion_utils): fix bugprone-narrowing-conversions warnings


* style(pre-commit): autofix

---------

+++ common/autoware_motion_utils/src/resample/resample.cpp
@@ -215,7 +215,8 @@ autoware_internal_planning_msgs::msg::PathWithLaneId resamplePath(
           resampling_arclength.at(j) = distance_to_resampling_point;
         } else {
           resampling_arclength.insert(
-            resampling_arclength.begin() + j, distance_to_resampling_point);
+            resampling_arclength.begin() + static_cast<std::ptrdiff_t>(j),
+            distance_to_resampling_point);
         }
         break;
       }
@@ -349,9 +350,9 @@ autoware_internal_planning_msgs::msg::PathWithLaneId resamplePath(
   for (size_t i = 0; i < resampled_path.points.size(); ++i) {
     autoware_planning_msgs::msg::PathPoint path_point;
     path_point.pose = interpolated_pose.at(i);
-    path_point.longitudinal_velocity_mps = interpolated_v_lon.at(i);
-    path_point.lateral_velocity_mps = interpolated_v_lat.at(i);
-    path_point.heading_rate_rps = interpolated_heading_rate.at(i);
+    path_point.longitudinal_velocity_mps = static_cast<float>(interpolated_v_lon.at(i));
+    path_point.lateral_velocity_mps = static_cast<float>(interpolated_v_lat.at(i));
+    path_point.heading_rate_rps = static_cast<float>(interpolated_heading_rate.at(i));
     path_point.is_final = interpolated_is_final.at(i);
     resampled_path.points.at(i).point = path_point;
     resampled_path.points.at(i).lane_ids = interpolated_lane_ids.at(i);
@@ -413,7 +414,9 @@ autoware_internal_planning_msgs::msg::PathWithLaneId resamplePath(
           } else if (dist_to_following_point < autoware::motion_utils::overlap_threshold) {
             resampling_arclength.at(i) = *distance_to_stop_point;
           } else {
-            resampling_arclength.insert(resampling_arclength.begin() + i, *distance_to_stop_point);
+            resampling_arclength.insert(
+              resampling_arclength.begin() + static_cast<std::ptrdiff_t>(i),
+              *distance_to_stop_point);
           }
           break;
         }
@@ -501,9 +504,9 @@ autoware_planning_msgs::msg::Path resamplePath(
   for (size_t i = 0; i < resampled_path.points.size(); ++i) {
     autoware_planning_msgs::msg::PathPoint path_point;
     path_point.pose = interpolated_pose.at(i);
-    path_point.longitudinal_velocity_mps = interpolated_v_lon.at(i);
-    path_point.lateral_velocity_mps = interpolated_v_lat.at(i);
-    path_point.heading_rate_rps = interpolated_heading_rate.at(i);
+    path_point.longitudinal_velocity_mps = static_cast<float>(interpolated_v_lon.at(i));
+    path_point.lateral_velocity_mps = static_cast<float>(interpolated_v_lat.at(i));
+    path_point.heading_rate_rps = static_cast<float>(interpolated_heading_rate.at(i));
     resampled_path.points.at(i) = path_point;
   }
 
@@ -556,7 +559,9 @@ autoware_planning_msgs::msg::Path resamplePath(
           } else if (dist_to_following_point < autoware::motion_utils::overlap_threshold) {
             resampling_arclength.at(i) = *distance_to_stop_point;
           } else {
-            resampling_arclength.insert(resampling_arclength.begin() + i, *distance_to_stop_point);
+            resampling_arclength.insert(
+              resampling_arclength.begin() + static_cast<std::ptrdiff_t>(i),
+              *distance_to_stop_point);
           }
           break;
         }
@@ -684,12 +689,12 @@ autoware_planning_msgs::msg::Trajectory resampleTrajectory(
   for (size_t i = 0; i < resampled_trajectory.points.size(); ++i) {
     autoware_planning_msgs::msg::TrajectoryPoint traj_point;
     traj_point.pose = interpolated_pose.at(i);
-    traj_point.longitudinal_velocity_mps = interpolated_v_lon.at(i);
-    traj_point.lateral_velocity_mps = interpolated_v_lat.at(i);
-    traj_point.heading_rate_rps = interpolated_heading_rate.at(i);
-    traj_point.acceleration_mps2 = interpolated_acceleration.at(i);
-    traj_point.front_wheel_angle_rad = interpolated_front_wheel_angle.at(i);
-    traj_point.rear_wheel_angle_rad = interpolated_rear_wheel_angle.at(i);
+    traj_point.longitudinal_velocity_mps = static_cast<float>(interpolated_v_lon.at(i));
+    traj_point.lateral_velocity_mps = static_cast<float>(interpolated_v_lat.at(i));
+    traj_point.heading_rate_rps = static_cast<float>(interpolated_heading_rate.at(i));
+    traj_point.acceleration_mps2 = static_cast<float>(interpolated_acceleration.at(i));
+    traj_point.front_wheel_angle_rad = static_cast<float>(interpolated_front_wheel_angle.at(i));
+    traj_point.rear_wheel_angle_rad = static_cast<float>(interpolated_rear_wheel_angle.at(i));
     traj_point.time_from_start = rclcpp::Duration::from_seconds(interpolated_time_from_start.at(i));
     resampled_trajectory.points.at(i) = traj_point;
   }
@@ -746,7 +751,9 @@ autoware_planning_msgs::msg::Trajectory resampleTrajectory(
           } else if (dist_to_following_point < autoware::motion_utils::overlap_threshold) {
             resampling_arclength.at(i) = *distance_to_stop_point;
           } else {
-            resampling_arclength.insert(resampling_arclength.begin() + i, *distance_to_stop_point);
+            resampling_arclength.insert(
+              resampling_arclength.begin() + static_cast<std::ptrdiff_t>(i),
+              *distance_to_stop_point);
           }
           break;
         }
